refactor: reservation 테이블이 schedule_id를 참조하도록 변경
- Reservation 객체에 Schedule 객체를 상태로 들고있게끔 변경
- ScheduleRepository 추가
- 변경으로 인한 테스트 깨짐 수정
- 인수테스트에서만 스프링컨텍스트를 띄우게 끔 변경
  - repository는 @JdbcTest + @Transactional
  - service는 @Transactional
- 테스트 전용 더미데이터 추가
  - 테스트데이터 연동을 위환 application-test.yml 추가
  - 각 테스트 프로필 설정
- dto에 요청검증 어노테이션 제거

+++ src/main/java/roomescape/reservation/application/ThemeService.java
@@ -3,6 +3,7 @@ package roomescape.reservation.application;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.reservation.domain.Theme;
+import roomescape.reservation.infra.ScheduleRepository;
 import roomescape.reservation.infra.ThemeRepository;
 import roomescape.reservation.presentation.dto.request.ThemeSaveRequest;
 import roomescape.reservation.presentation.dto.response.ThemeFindResponse;
@@ -15,6 +16,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ThemeService {
     private final ThemeRepository themeRepository;
+    private final ScheduleRepository scheduleRepository;
 
     public ThemeSaveResponse save(ThemeSaveRequest body) {
         Theme newTheme = themeRepository.save(body.toDomain());
@@ -22,6 +24,9 @@ public class ThemeService {
     }
 
     public void delete(long id) {
+        if (scheduleRepository.existsByThemeId(id)) {
+            throw new IllegalStateException("themeId= " + id + " 인 테마를 사용하는 스케줄이 있어 삭제할 수 없습니다.");
+        }
         themeRepository.deleteById(id);
     }
 
@@ -35,7 +40,7 @@ public class ThemeService {
         return ThemeFindResponse.of(themes);
     }
 
-    public List<ThemeFindResponse> findAll(){
+    public List<ThemeFindResponse> findAll() {
         List<Theme> themes = themeRepository.findAll();
         return ThemeFindResponse.of(themes);
     }
